requirements.txt
# Required libraries for the project
requests
pandas
pandas-gbq
google-auth
google-auth-oauthlib